+++ internal/graph/cached_resolver_test.go
@@ -450,7 +450,6 @@ func TestResolveCheckFromCache(t *testing.T) {
 	}
 
 	for _, test := range tests {
-		test := test
 		t.Run(test.name, func(t *testing.T) {
 			t.Parallel()