+++ pkg/server/test/list_objects.go
@@ -468,7 +468,6 @@ func TestListObjects(t *testing.T, ds storage.OpenFGADatastore) {
 	}
 
 	for _, test := range testCases {
-		test := test
 		t.Run(test.name, func(t *testing.T) {
 			t.Parallel()
 			ctx := context.Background()